Rabin Karp Algorithm

+++ edu/concepts/string/StringHashing.java
@@ -8,14 +8,19 @@ public class StringHashing {
         System.out.print("Enter the string.. ");
         Scanner sc = new Scanner(System.in);
         String str = sc.nextLine();
-        int prime = 53;
-        int mod = 1000000007;
+
+        /*
+        * prime should be a prime number and
+        * greater than the character set.
+        * */
+        long prime = 53;
+        long mod = 1000000007;
         System.out.println(new StringHashing().getHash(str, prime, mod));
     }
 
-    private int getHash(String str, int prime, int mod){
-        int hash = 0;
-        int pow = 1;
+    private long getHash(String str, long prime, long mod){
+        long hash = 0;
+        long pow = 1;
 
         for (int i = 0; i < str.length(); i++) {
             hash += (((int)str.charAt(i) - 'a' + 1) * pow) % mod;
